Refactor DeleteTokenPage
