+++ client/src/pages/DeleteTokenPage/components/DeleteTokenItem.tsx
@@ -2,11 +2,11 @@ import React from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import classnames from 'classnames';
 
-import TokenBadge from '../../components/Badges/TokenBadge/TokenBadge';
+import TokenBadge from '../../../components/Badges/TokenBadge/TokenBadge';
 
-import { CheckedTokenType } from './types';
+import { CheckedTokenType } from '../types';
 
-import classes from './styles/DeleteToken.module.scss';
+import classes from '../styles/DeleteTokenItem.module.scss';
 
 type Props = {
   hasLowBalance: boolean;
