+++ client/src/pages/DeleteTokenPage/styles/DeleteTokenItem.module.scss
@@ -1,0 +1,14 @@
+@import '../../../assets/styles/colors.scss';
+
+.checkIcon {
+  color: $dark-jungle-green;
+  cursor: pointer;
+  font-size: 20px;
+  margin-left: 40px;
+  justify-self: flex-end;
+
+  &.inactiveIcon {
+    color: $manatee;
+    cursor: not-allowed;
+  }
+}
